use 64-bit indexers also in windows

+++ ctpfrec/cy.pyx
@@ -12,8 +12,8 @@ import time
 ## Note: As of the end of 2018, MSVC is still stuck with OpenMP 2.0 (released 2002), which does not support
 ## parallel for loops with unsigend iterators. If you are using a different compiler, this part can be safely removed
 IF UNAME_SYSNAME == "Windows":
-	obj_ind_type = ctypes.c_long
-	ctypedef long ind_type
+	obj_ind_type = ctypes.c_longlong
+	ctypedef long long ind_type
 	ctypedef double long_double_type
 	obj_long_double_type = ctypes.c_double
 ELSE:
